requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dateparser==1.1.7
dnspython==2.3.0
fastapi==0.92.0
fastapi-utils==0.2.1
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
motor==3.1.1
multidict==6.0.4
pycodestyle==2.10.0
pydantic==1.10.4
PyJWT==2.6.0
pymongo==4.3.3
python-binance==1.0.16
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.25.0
twilio==7.16.3
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.2
